refactor(common): 공통 모듈 폴더 구조 수정
