+++ backend-server/src/main/java/shop/bluebooktle/backend/book/service/Impl/BookLikesServiceImpl.java
@@ -15,9 +15,9 @@ import shop.bluebooktle.backend.book.repository.BookRepository;
 import shop.bluebooktle.backend.book.service.BookLikesService;
 import shop.bluebooktle.backend.user.repository.UserRepository;
 import shop.bluebooktle.common.entity.User;
-import shop.bluebooktle.common.exception.BookLikesAlreadyChecked;
-import shop.bluebooktle.common.exception.BookNotFoundException;
 import shop.bluebooktle.common.exception.UserNotFoundException;
+import shop.bluebooktle.common.exception.book.BookLikesAlreadyChecked;
+import shop.bluebooktle.common.exception.book.BookNotFoundException;
 
 @Service
 @RequiredArgsConstructor
