+++ common-module/src/main/java/shop/bluebooktle/common/exception/book/BookLikesAlreadyChecked.java
@@ -1,4 +1,4 @@
-package shop.bluebooktle.common.exception;
+package shop.bluebooktle.common.exception.book;
 
 public class BookLikesAlreadyChecked extends RuntimeException {
 	public BookLikesAlreadyChecked(String message) {
